Prefix control plane URLs, removed stats, log fatals
Control plane URLs, i.e. the helper URLs that we give you for free such as
/heartbeat, all now have a prefix of a single underscore _ , i.e. /_heartbeat

The stats package has been removed in anticipation of adding CloudFlare metrics
that are more meaningful for our applications.

The negroni.Run() wrapper has been removed so that we use our logging interface
for Fatal errors, and can format that consistently for our logging tools.